+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/Screen.kt
@@ -55,6 +55,7 @@ sealed class Screen(open val title: StringResource, open val icon: DrawableResou
 //        Res.drawable.key)
     data class MCPAccessKey(
         val accessKey: AIPortToolAccessKey?=null,
+        val integrationGuide: Boolean = false,
         val dialog: MCPKeyDialog = MCPKeyDialog.None
     ):Screen(Res.string.mcp_keys,Res.drawable.key)