+++ composeApp/src/jvmMain/kotlin/ai/mcpdirect/mcpdirectstudioapp/main.kt
@@ -217,6 +217,7 @@ fun main() = application {
 //                                    )
                                 MCPdirectKeyScreen(
                                     screen.accessKey,
+                                    screen.integrationGuide,
                                     paddingValues
                                 )
                             }